+++ airflow/contrib/example_dags/example_emr_job_flow_automatic_steps.py
@@ -22,8 +22,8 @@ This is an example dag for a AWS EMR Pipeline with auto steps.
 from datetime import timedelta
 
 from airflow import DAG
-from airflow.contrib.sensors.emr_job_flow_sensor import EmrJobFlowSensor
 from airflow.providers.amazon.aws.operators.emr_create_job_flow import EmrCreateJobFlowOperator
+from airflow.providers.amazon.aws.sensors.emr_job_flow import EmrJobFlowSensor
 from airflow.utils.dates import days_ago
 
 DEFAULT_ARGS = {
