+++ airflow/contrib/sensors/emr_job_flow_sensor.py
@@ -16,48 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from airflow.providers.amazon.aws.hooks.emr import EmrHook
-from airflow.providers.amazon.aws.sensors.emr_base import EmrBaseSensor
-from airflow.utils.decorators import apply_defaults
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.emr_job_flow`."""
 
+import warnings
 
-class EmrJobFlowSensor(EmrBaseSensor):
-    """
-    Asks for the state of the JobFlow until it reaches a terminal state.
-    If it fails the sensor errors, failing the task.
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.sensors.emr_job_flow import EmrJobFlowSensor  # noqa
 
-    :param job_flow_id: job_flow_id to check the state of
-    :type job_flow_id: str
-    """
-
-    NON_TERMINAL_STATES = ['STARTING', 'BOOTSTRAPPING', 'RUNNING',
-                           'WAITING', 'TERMINATING']
-    FAILED_STATE = ['TERMINATED_WITH_ERRORS']
-    template_fields = ['job_flow_id']
-    template_ext = ()
-
-    @apply_defaults
-    def __init__(self,
-                 job_flow_id,
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.job_flow_id = job_flow_id
-
-    def get_emr_response(self):
-        emr = EmrHook(aws_conn_id=self.aws_conn_id).get_conn()
-
-        self.log.info('Poking cluster %s', self.job_flow_id)
-        return emr.describe_cluster(ClusterId=self.job_flow_id)
-
-    @staticmethod
-    def state_from_response(response):
-        return response['Cluster']['Status']['State']
-
-    @staticmethod
-    def failure_message_from_response(response):
-        state_change_reason = response['Cluster']['Status'].get('StateChangeReason')
-        if state_change_reason:
-            return 'for code: {} with message {}'.format(state_change_reason.get('Code', 'No code'),
-                                                         state_change_reason.get('Message', 'Unknown'))
-        return None
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.emr_job_flow`.",
+    DeprecationWarning, stacklevel=2
+)
