+++ docs/operators-and-hooks-ref.rst
@@ -356,7 +356,7 @@ These integrations allow you to perform various operations within the Amazon Web
        :mod:`airflow.providers.amazon.aws.operators.emr_create_job_flow`,
        :mod:`airflow.providers.amazon.aws.operators.emr_terminate_job_flow`
      - :mod:`airflow.providers.amazon.aws.sensors.emr_base`,
-       :mod:`airflow.contrib.sensors.emr_job_flow_sensor`,
+       :mod:`airflow.providers.amazon.aws.sensors.emr_job_flow`,
        :mod:`airflow.contrib.sensors.emr_step_sensor`
 
    * - `Amazon Kinesis Data Firehose <https://aws.amazon.com/kinesis/data-firehose/>`__
