+++ tests/providers/amazon/aws/sensors/test_emr_job_flow.py
@@ -24,7 +24,7 @@ from unittest.mock import MagicMock, patch
 from dateutil.tz import tzlocal
 
 from airflow import AirflowException
-from airflow.contrib.sensors.emr_job_flow_sensor import EmrJobFlowSensor
+from airflow.providers.amazon.aws.sensors.emr_job_flow import EmrJobFlowSensor
 
 DESCRIBE_CLUSTER_RUNNING_RETURN = {
     'Cluster': {
